requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-jazzmin==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
gunicorn==22.0.0
inflection==0.5.1
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==70.0.0
sqlparse==0.5.0
uritemplate==4.1.1
